refactor(경로): dto 패키지 생성 및 적용
